[Test] Improved test suite * Refactoring of RptDisp * Modified specification for levels
