+++ .circleci/config.yml
@@ -12,7 +12,8 @@ test: &test
     - run: make build-python
     - run: source .circleci/install_tvm.sh
     - run: pip install --user pytest 
-    - run: pip install --user future 
+    - run: pip install --user future
+    - run: pip install --user imageio
     - run: python -m pytest tests
     - run: pip install --user mxnet
     - run: python -m pytest samples
