+++ python/heterocl/__init__.py
@@ -9,6 +9,5 @@ from .debug import hcl_excepthook
 from .tvm.intrin import *
 from .tvm.stmt import Partition
 from .tvm.expr import IO
-from .rptdisp import *
 
 sys.excepthook = hcl_excepthook
