+++ python/setup.py
@@ -46,6 +46,5 @@ setup(
       'tabulate',
       'sodac',
       'pandas',
-      'imageio',
       ],
   **setup_kwargs)
