+++ tests/test_report_data/expected.json
@@ -3,5 +3,7 @@
   "LoopQuery" : ["Trip Count, Latency Iteration Latency Pipeline II Pipeline Depth", "P_x, 412, 8643760, 20980, N/A, N/A", "+ P_y, 617, 20978, 34, N/A, N/A", "++ P_z, 3, 15, 5, N/A, N/A", "A_x1, 412, 4068088, 9874, N/A, N/A", "+ A_y1, 617, 9872, 16, N/A, N/A"],
   "ColumnQuery" : ["Latency", "P_x, 8643760", "+ P_y, 20978", "++ P_z, 15", "A_x1, 4068088", "+ A_y1, 9872", "X_x_reuse_X_y_reuse, 29741875", "Y_x_reuse1_Y_y_reuse1, 29741875", "R_x3_R_y2, 252173", "F_x4_F_y3, 252156"],
   "LevelQuery" : ["Trip Count, Latency Iteration Latency Pipeline II Pipeline Depth", "P_x, 412, 8643760, 20980, N/A, N/A", "+ P_y, 617, 20978, 34, N/A, N/A", "A_x1, 412, 4068088, 9874, N/A, N/A", "+ A_y1, 617, 9872, 16, N/A, N/A", "X_x_reuse_X_y_reuse, 254204, 29741875, N/A, 117, 125", "Y_x_reuse1_Y_y_reuse1, 254204, 29741875, N/A, 117, 125", "R_x3_R_y2, 252150, 252173, N/A, 1, 25", "F_x4_F_y3, 252150, 252156, N/A, 1, 8"],
-  "MultiQuery" : ["Trip Count, Latency Iteration Latency Pipeline II Pipeline Depth", "P_x, 412, 8643760, 20980, N/A, N/A", "+ P_y, 617, 20978, 34, N/A, N/A", "A_x1, 412, 4068088, 9874, N/A, N/A", "+ A_y1, 617, 9872, 16, N/A, N/A"]
+  "LevelQueryOOB" : ["Trip Count, Latency Iteration Latency Pipeline II Pipeline Depth", "P_x, 412, 8643760, 20980, N/A, N/A", "+ P_y, 617, 20978, 34, N/A, N/A", "++ P_z, 3, 15, 5, N/A, N/A", "A_x1, 412, 4068088, 9874, N/A, N/A", "+ A_y1, 617, 9872, 16, N/A, N/A", "X_x_reuse_X_y_reuse, 254204, 29741875, N/A, 117, 125", "Y_x_reuse1_Y_y_reuse1, 254204, 29741875, N/A, 117, 125", "R_x3_R_y2, 252150, 252173, N/A, 1, 25", "F_x4_F_y3, 252150, 252156, N/A, 1, 8"],
+  "MultiQuery" : ["Trip Count, Latency Iteration Latency Pipeline II Pipeline Depth", "P_x, 412, 8643760, 20980, N/A, N/A", "+ P_y, 617, 20978, 34, N/A, N/A", "A_x1, 412, 4068088, 9874, N/A, N/A", "+ A_y1, 617, 9872, 16, N/A, N/A"],
+  "AllQuery" : ["Latency", "P_x, 8643760", "+ P_y, 20978", "A_x1, 4068088", "+ A_y1, 9872"]
 }
